Active Trigger Changes
Added a new property to the InstanceModel to store the count of active property triggers.

Altered the ActiveTriggers property in the InstanceModel to store the count of active contatc triggers.

Altered the GetActiveTriggers method in InstanceConfig to get both the contact and property trigger counts.

Changed GetRAGStatus method to look at the correct active triggers property.

Changed OutputResult method to output the active trigger data to the new field and the renamed field.